backend/requirements.txt
bcrypt==3.1.3
crossbar==17.5.1
mysqlclient==1.3.11
python-gnupg==0.4.1
redis==2.10.5
SQLAlchemy==1.1.6
tornado==4.5.2
virtualenv==15.1.0
